16 bits are enough here, save bytes :-)
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@14270 a1c6a512-1295-4272-9138-f99709370657

+++ apps/gui/gwps-common.c
@@ -144,7 +144,7 @@ bool update_onvol_change(struct gui_wps * gwps)
 
 bool ffwd_rew(int button)
 {
-    static const int ff_rew_steps[] = {
+    static const uint16_t ff_rew_steps[] = {
       1000, 2000, 3000, 4000,
       5000, 6000, 8000, 10000,
       15000, 20000, 25000, 30000,
